+++ app/models/conversion/voluntary/tasks/conversion_grant.rb
@@ -1,4 +1,4 @@
-class Conversion::Voluntary::Tasks::ConversionGrant < TaskList::Task
+class Conversion::Voluntary::Tasks::ConversionGrant < TaskList::OptionalTask
   attribute :check_vendor_account
   attribute :payment_form
   attribute :send_information